requirements.txt
bytez
pymongo
python-dotenv